app/requirements.txt
django-haystack==2.8.1
Django==2.1.7
elasticsearch>=2.0.0,<3.0.0
psycopg2-binary==2.7.7
